[[Add search and fix continue]]
* Added a "search" action, which searches labels and unit names
(unit::description()) for the string you type. Case sensitive. Triggered by the
"/" key. Hitting enter at the enter-search-term dialog repeats the
previous search. Searches labels first, then units.

* Fixed a problem with the conditions for displaying the "continue move" action
in the context menu (it was checking things properly in
turn_info::continue_move(), but not in can_execute_command(...)). This caused
"Continue Move" to appear to be a valid option, but clicking it would have no
affect.