Added a "shared vision" option to multiplayer games...
...(and single-player games, via the scenario cfg file). It causes all
allied players to be able to see the same fog and shroud.

Notable changes by file:

* data/translations/english.cfg = added a "Shared Vision" string

* src/multiplayer.[ch]pp = added a "Shared Vision" checkbox

* src/multiplayer_connect.[ch]pp = added a shared_vision parameter to
load_game

* src/playlevel.cpp = clear shroud for each team once up front, before
the first player's turn

* src/replay.cpp = fixed several commands not clearing fog/shroud
or refreshing the minimap properly

* src/team.[ch]pp = changed shrouded(x,y) and fogged(x,y) methods to
return true only if the square is shrouded/fogged for every allied
player, if and only if shared vision is turned on for this team.