Restructured the Editor UI:
Two terrains may now be selected in the palette, one is used for
filling when moving selections and cutting/pasting and will be used
when drawing with the right button.

The context menu is removed, instead the function of the left button
is chosen in the right panel and operations that do not need a
location to work is chosen from an additional menu.