Made the editor use the default image for the terrains...
...while drawing them so that updates are visible immediately, but
things do not have to be rebuilt. (Fixes some CVS-only issues)

Changed rebuild_terrain() to set the image for that location to the
default image of the terrain. It would still be nice to be able to
rebuild locations on the fly, but if it really is too complicated,
this solution is not that bad.