A few changes to the network threading code.
Don't rely too much on iterator validity property of associative
containers; in fact, remove them and use a vector of pointed buffers
to increase performance. Correctly use conditions: a mutex should not
be freed just after it has been acquired through a condition wait; and
don't send a condition with the mutex being hold (it kind of defeats
the purpose of a condition). These changes require testing!