First step in improving the help system speed.
Topics (unit topics only for the time being) are now created on the
fly; they are generated by virtual function objects, owned by pseudo
shared pointers, and their value is cached once computed (I agree,
it's a bit ugly). The wrong code identation is intended: it
considerably reduces the size of the patch.