At first it was supposed to be a small patch...
to fix weapons not being translated in the attack box. But now it also
handles the way the battle statistics are stored in memory: the
optional part is now strictly user-oriented, and hence can be fully
translated. As a consequence, the common part is much smaller, almost
a POD; it should speed up the AI since both the memory footprint and
the _struction time are notably reduced. To avoid changing too much
code, there are still two strings in the common structure in order to
handle attack specials; they could be replaced later on by an integer
bitset to further speed up the game.