Revised images for footpads and outlaws...
...which fix the lack of alpha shading and defense marks, which change
the hair color and style of the "dirty mullet" formerly worn by the
footpad to something more compatible with a female unit which is
supposed to look "pretty".  Also added images for a female version of
the footpad, and made a new version of my former hack of a female
outlaw, which is more consistent with the male outlaw.