Remove the line_source stuff...
...and embed the inclusion data directly into the stream. Now we are
finally freed of any stateful WML processing, and we can go on with
the preprocessor cleanup. Small drawback: the error display is
temporarily disabled in the parser; I will add it back in a few
days. In the meantime, the commit will serve to detect for situations
where the inclusion data is embedded in the wrong place.