Fix for bug #10704;
Before saving and loading we check if the save game name contains any
invallid characters like '?'. At loading time the user is warned to
save the map under an other name and at saving time the user gets a
warning and won't be able to save the map with illegal characters in it.