Fix GCC warning about dubious copy constructor...
...by explicitly specifying that the base class is not
copied. Speaking about the copy constructor, because the
default_defines is not copied and the pointer is not updated, it would
lead to a use-after-free memory access if the buffer was to be
destroyed. I can't think of any normal situation where this would
happen, but better safe than sorry. So let's push the ownership one
level upward. As a nice side-effect, the preproc_map is no more
constructed/destructed when useless.