I changed the code that handles campaign_list...
...to prepare for being able to return information about a subset of
the campaigns. I save a pointer to the request WML. I start with no
campaigns and copy them in one by instead of the old way. I left the
passphrase blanking to a later loop that runs after the list of
campaigns are built. There is probably a more efficient way to do it
when I copy over each individual campaign summary, but it seems easier
not to mess up leaving it work like it did previously.  My next task
will be to handle conditions, but I wanted to get the base code
checked in after testing, before I started adding in the new tests.