I changed weblist.pl and webtgz.pl to display the standard wesnoth logo banner
(except for when webtgz.pl successfully returns a tgz file).
I also changed webtgz.pl to only request campaign info for the campaign of
interest. In the future (after getting campaign summary info loaded for old
campaigns on the campaign server) this request won't be needed at all and will
be removed.