Because the translation heuristic is imperfect...
...it will sometimes pick up the same language twice for a given
campaign. (Currently Invasion of Eliador has this issue with sv.) This
isn't a problem on the campaign server side.  However it looks bad on
the web page. So I now have the campaign list suppressing duplicate
langauges.

I also fixed a bug where the languages weren't sorted the intended way.