prkill 1.10, helper script to calculate probabilities to kill in a skirmish.
This is more general than the in-game code in Damage Calculations
since it handles berserk.  It computes the joint conditional
probability function of the hitpoints each unit will have at the end
of the skirmish.  The algorithm is currently slow: O(r*(n+N)*(h*H))
operations (n and N are number of swings per round, h and H are
hitpoints, and r is 30 if berserk is in effect otherwise 1).  Hence,
probably not yet suitable for AI use.