Change to backstab...
...to work when the unit opposite the attacker is an enemy of the
defender and is not incapacitated. (Note someone familiar with the
code style might want to review this to see if I broke any style
guidelines.  The main issue would probably be how I got access to
teams_ and units_.)

Updated the builtin help to describe backstab consistant with the above and
mentioned the 30 round limit in the beserk description.

Renamed "Across The River" as "Across the River" as articles don't seem
to be capitalized in most maps.

Renamed the config file containing "Across the River" to be consistant with
other maps, so it appears in a more normal order in the displayed list.

Changed "Load Game ..." to be the first multiplayer game option to make this
option easier to notice.