I tweaked how experience_modify for units types worked.
Previously there was an odd check to see if it was still 100 and if so
to divide by 100 and multiply by the new value. This essentially had
the first setting override any further settings. I changed this to
just use the value specified, so the last setting takes
precedence. For experience needed, I changed it to round and put in a
mimimum of 1.