Factorized core source files into a libwesnoth-core library.
This library is intended for non-graphical stuff, and suitable for
server-type apps.  It is only a convenience library for build purposes
at this time, but since I took care of only including the those core
modules that are clean dependency-wise, and make an independant set of
files, we could even (technically speaking) make it a shared lib right
now.  More work is needed to move other files in this independant
core, and to create a graphics-core library as well.