Fix http://gna.org/bugs/index.php?func=detailitem&item_id=4996
When calculating paths, we need to differentiate between the team the
unit is on, and the team doing the calculation.  These are *different*
for Show Enemy Moves and highlighting enemy units' range: we *know*
they can't walk through our ambush.

In general, checking map.count(loc) != 0 is not the right thing to do,
you should use find_visible_unit().  A player can currently find an
invisible unit on your keep by trying to recall on that hex, for
example, but I consider that too minor to worry about, and difficult
to solve.