Fix non-US keyboards:
https://gna.org/bugs/index.php?func=detailitem&item_id=4380

This is non-trivial.  We don't want to break people's customized
hotkeys, yet we need to change the fundamental way keyhanding works.
Above all, we do not have a keyboard map.

SDL gives the unicode value of the keys pressed, as well as the key
itself and the modifiers.  We need to use this unicode value sometimes
(eg. for ':', which is not shift-; on non-US keyboards) and the
keycode at other times (F1 has no unicode value, and shift-space is
the same as space).  We do this as follows:

(1) Single-letter keys get treated as the character,

(2) Multi-letter keys (function keys, delete, etc) and space get
treated as the actual key.

However, we want backwards compatibility, so *if* they specify a shift
modifier we fall back to old-style keycode.

Finally, if the control key is pressed we often need to map the
unicode value back to the "raw" value to figure out what the character
is.  We only do this if the character value is less than 64: it's
probably not a universal rule.