Fix AI skirmisher landing on same tile (bug #5214).
In last episode (2006-01-19T09:50:19Z!rusty@rustcorp.com.au fixing bug #4191) our hero had removed the
check in ai.cpp's move_unit_partial which stopped a skirmisher from
triggering an ambush.

Unfortunately, this now meant that the AI move stopped on passing by
*any* enemy unit, in the assumption it was ambushed.  Not true: AI
unit could be a skirmisher.  This resulted in aborting a move while
over our own unit: this cannot happen by ambush because an invisible
unit cannot be adjacent one of our units.

This introduces the correct fix: ONLY abort the move if the adjacent
enemy unit has a ZOC and is actually invisible, otherwise continue (in
which case the AI unit must be a skirmisher, or the AI should not have
tried such a move).