Fix OOS bug...
...where one player (or observer) chooses a different defensive
weapon, as seen by Soliton (IIRC)

Xan pointed out this was a possibility: we calculate defense weapon
now based on attack_prediction code, but this uses floating point
calculations, which means that in the case where results are close,
rounding differences can mean that both ends choose different weapons.

The answer is to store the defense weapon as well as the attack
weapon, and pass it to all the relevent routines.