Infrastructure to fix slow auto-saves.
We currently save the game by creating a huge config containing all
the details, then writing it out.  We spend most of our time copying
configs around, causing autosaves to be very slow.

This introduces a config_writer class for writing out configs in
pieces, and exposes more of the parser's writing code for it to use.

Note: does not yet do binary saves.