Fix bug 6168 (access beyond end of array when AI calculates vs. drain).
When an unit doesn't drain, we only allocate space up to current HP
(since we know it won't increase).

So when the AI simulates a series of attacks against a draining
defender, if the first attack is against the non-draining weapon, the
summary is ranges from 0 to orig_defender_hp.  If the second attack is
against the draining weapon, we allocate 0 to max_defender_hp, and
then try to copy across the old results, which are smaller.