Changed the custom lava terrain in UtbS to use the new lava transitions.
I couldn't get it working when adding the tranistions in the campaign
terrain_graphics.cfg so I added it to the normal terrain_graphics.cfg
and put it between ifdef's. Since this will be included in the 1.2
release it would be nice if some more eyes would look at this.  (The
problem with the campaign terrain_graphics.cfg was that the
transitions of the stone path were displayed in the air, these
transitions shouldn't be there in the first place. And some more
simular problems.)