Hardcoded 5 buttons per row, needs to be depending on the theme.
Added the code to let buttons stay in a pressed state this requires
the buttons to be of the type gui::button::TYPE_CHECKED but that
give glitches so kept them as gui::button::TYPE_PRESS which doesn't
allowed the pressed code to work. Once the glitches have been 
sorted out it should work with staying pressed down.