made an option to set the number of fps that should be shown,
...not all code honours this yet

the default is 50 fps

changed the draw delay logic it now determines when the next draw
should take place, this avoids adding a delay when the code is already
lagging. That problem occurs with scrolling and animations. The engine
now stays much closer to the desired fps, provided that the system it
runs on is fast enough.