Wiring in the 'Ruffian' to tSG.
This L0 Unit appears in the easy and medium difficulties in the first
three scenarios.  The Ruffian's stats may need adjustment, as is they
were taken directly from the XE henchman.  The hope is that this unit
will be useful for other campaigns, but i designed him for tSG.
Rather than radically change his concept, the unit should be moved (or
duplicated) to the tSG folders.  Also created a smaller club and gave
it to the Footpad as well.