Get rid of the Garak special units.
This involved giving Garak an object to modify his ranged attack and
creating a global event to re-hack his portrait if he adances to
Desert Marshal (his stats as a Marshal are stock).

The portrait-setting on advancement is the same thing that's done with Elyssa.
I've abstracted this into a SET_PORTRAIT_ON_ADVANCE macro.

Also, tweak Desert Captain to use SPECIAL_NOTES_LEADERSHIP.