Correct the fix for bug #9191.
Update the sidebar after a recruit only if it happens under the mouse.
The previous fix caused a quick highlight of the hex, which was needed
by the bad current sidebar code.

Also I'm now less sure that we want to always display info about the
enemy recruiting.

And there is the little question: the acting unit is the recruiter or
the recruited?