Important optimization of the cache size and zooming...
...for standard terrain at default zoom level:

Previously the image type UNMASKED was a cache for hex-masking and
scaling.  added a type HEXED for caching the hex-masking, when using
it skip the UNSCALED, since these hexed images will normally never be
used non-hexed.

So this new cache doesn't cost memory and hex-masking doesn't need to
be redone when zooming.

At default zoom level, we can now skip the UNMASKED cache, so will
spare a cache level in memory

In theory, we win 25%-33% of memory space (depend of the use of
BRIGHTENED) for standard terrain hexes.

Note: the hex-masking and scaling operations are now inverted, it
doesn't seems to have a visible effect.