Eliminate all but three compiler warnings.
Most are banished by use of a new UNUSED macro that expnds to the
right GNU C++ attribute but an empty string elsewhere.  I have the
definition being generated into config.h, which therefore has to be
included a few places it wasn't before.  There might be a better place
to put it.