Experiment to show the direction from the attack.
This is for helping click accurancy, specially because of the new
"attack a nearby unit from an other direction" feature.

Please test and comment the usefullness and look of this. Maybe add a
"bad" highlight for blocked directions ?  The code and images are just
a quick hack to see if and how we want this.