New experiment to choose the attack direction:
Use the last highlighted empty hex to choose the direction of the attack.

This is a user-test and you will need a short(I think) adaptation time.

It's mainly to fix accuracy problem with the old system (exposed more
with the new possibility to choose the direction at close range)