Make random_traits = "yes" the default.
This makes the feature more powerful as well. You can define less than
the max number of traits manually and have the remainder filled in
randomly. It also forces undead and mechanical to always be assigned
when available for a unit.

Currently no random traits are assigned to leaders. The problem is
that random number generation triggered by the leader creation isn't
properly handled in MP games, so only must have traits can be given
out currently.  This is controlled by [trait] configuration, so down
the road traits that are not unbalancing for MP play (quick was
especially seen as a problem for MP) may be given to leaders.