Fix a little bug: undo a movement starting from an impassable terrain
(which is possible) was very slow (99 times slower), need ~15sec.

Note also that a unit with a lot of MP but high movement cost have the same
problem. So now use 10 as maximum MP cost for the animation speed.
Still very slow (~1sec/tile), but you can use acceleration if it's frequent.