Address bug #9874.
With this change, loading a save made in linger mode should put you in
linger mode again, with no unit movement available and recall
disabled.

I believe eecall being disabled in linger mode is an efect of the
browse flag being on; this has nothing to do with linger mode per se.
Please file "enable read-only view of the recall list in browse mode"
as a separate FR, as it has other implications.