Add a minor tributary river passing near the human player's keep...
...in the north part of the map. This adds some visual interest, but
more importantly it actually makes mermen tactically useful in this
scenario -- the player can quickly send them to scout to the SE, where
a number of interesting features are visible from the water.

This change is pursuant to some forum complaints about mermen being useless
in most of this campaign.