In help, allow to hide a page from the left-panel view.
Plan to use it for hiding the "Unknown Unit" page or other error/info pages

Simply use the linux convention: topic id starting with a "." are hidden.

Note that open a hidden topic in a section, will also open the section.
Can be useful, but if not wanted, simply place the hidden topic at toplevel