Insert a new scenario, "Return To Kerlath",
...after "Into The Depths" and before "Vengeance" on the elf line of
tSG.  This is intended to be a relatively light scenario to smooth the
otherwise jarring direct transition from Deoran & Co entering Mal
M'Brin's chamber to being back at Westin for the final battle of the
campaign.

This isn't balance-tested yet.  It's patterned on TB:The Chase, with similar
terrain and start conditions, but a somewhat more dangerous opponent.

I've committed it so zookeeper and Mythological can get a look at it and
tune as needed.