Major retuning of starting villages in mainline.
One aspect of the village-income system that has always griped me is
that it gives players an incentive to run units away from the fighting
in order to grab villages that logically in storyline terms should
have started under their control to begin with.  Seizing the enemy's
villages is strategy; having to seize villages trivially near your own
keep (especially *behind* it) is just a pain in the ass.

This revision fixes that by adding STARTING_VILLAGES macros in a whole
boatload of places.  In most cases (a) these are villages the player
would have seized in turn 1 anyway, (b) the amount of additional
income per turn is small, and (c) I give the bad guys villages too for
the same startup acceleration.  The net effect of these changes should
be to get the scenarios off to a faster start without significantly
messing with scenario or campaign balance.

In a few cases the amount of additional income looks unbalancing and I
have adjusted minimum starting gold correspondingly.  That is probably
sufficient for campaign-starting scenarios, but the income increment
in mid-campaign scenarios (especially those with no minimum gold,
where the player is expected to start wealthy) may have a balance
effect on later scenarios that's harder to control.

I suggest campaign maintainers should re-evaluate the following scenarios for 
balance effects:

Eastern Invasion::The Outpost
Scepter of Fire::Closing The Gates
Scepter of Fire::Towards the Caves
Scepter of Fire::Searching For The Runecrafter
Heir To The Throne::The Seige Of Elensefar

I've also ensured that STARTING_VILLAGE calls within prestart and start 
events are moved out of them.  In general I've moved STARTING_VILLAGES 
calls to right after their side declarations.