Magic-number elimination
Replace the harcoded numeric constant '9' with gamemap::MAX_PLAYERS
where appropriate. Also replace START_POSITIONS, This doesn't
completely fix bug #6651, but it's a good running start.  This patch
does only the simple things, so as not to risk breakage before 1.3.10;
it marks two slightly trickier things that need to be changed with
FIXMES