[[Fix stall on wescamp upload.]]
Since the servers mainloop is single threaded (instead of multi as earlier 
assumed) the wescamp upload script blocks the entire server. To avoid this 
problem the script is started and the execution results are ignored which 
avoids the blocking. This is not very robust but the entire server will be
rewritten in the future so it's not that important. The forked execution
also lets two threads write to the std::cerr but since nobody read this it's
also no problem.