For the footpath when teleporting,
...use a halo image instead of footsteps in funky directions (but
continue to draw them for entering and leaving the villages). It's
mainly a visual indicator to show that there is something special
happening and that your footpath is not just stuck in the village.

It's a quick Gimp, but if/when replaced, it must stay light. Because
it's coded to stay visible like footsteps when the unit move and
teleport, and so it must work well under the teleport anim. If really
needed, we can add a new unit key specifing the image to use. Btw, the
siver mage seems to not have his teleport anim in cfg but the images
are there.