Simplify and clarify a little the map_label search function
...because patch #849: "iterator for labels_ should be initialized to
something" indicated that the code was not nice, even if it was
correct.  The patch pointed highlighted the bad code style but don't
solved the problem.  Also add a function allowing to get public label
hidden under a team only-label Maybe use it in the not-team-only field
of the dialog for this rare case?