Our hex-masking function doesn't support image different of 72x72
Not a problem, you can crop it or add tranparent parts.

But the previous bad result was very weird, and so hard to see your
error Now don't mask these image (you will see that bad parts are not
in the hex) Also add a warning to the standard output explaining what
to do.