Address bug #10419.
Having reviewed the way it's used, I don't believe using it will make Wesnoth a 
derivative work of boost.m4, so v3 terms will not propagate to the 
rest of the code.  (Obligatory note: I am not a lawyer.)

This is an issue we'll face again when and if autotools goes v3.