Updated the network code to handle gzipped data instead of binary wml.
Nobody sends it yet, but the server can receive it, once the server
has been restarted there will be another commit to make the clients
send gzipped data to the server.

There's a config variable to change the server to send gzipped data as
well. On the trunk server this switch can be set once everybody had
some time to update their clients.

Once 1.3.12 is no longer allowed on the server this can also be set on
the 1.3 server. Once it's no longer allowed there the server will be
modified to only send gzipped data and the switch removed.

(The campaign server won't be modified at this moment.)

NOTE *campaign server and server need to be included
serialization/binary_or_text.cpp*