Move hardcoded definition of the footpath-teleport images in [game-config]
Still think about allow ability to change it, but I prefer cache the
footpath before adding to much things to the real-time drawing part.

And it's also good to have a default images (works for any
discontinuous path) It will maybe be used if terrain teleport is added
one day.