Add a slighlty dark background (black with alpha 0.3) to hp/xp bars.
This helps bug #11030 (lower the risk of confusion with colors under
the bars, caused by terrains or by the unit, like the drake burner's
wing or the elvish fighter'sword (who always look near AMLA). Also
useful on snow terrains.  Maybe the 0.3 value must be tuned. This also
compensate a little the change of the alpha of bars from the previous
commit (brighter bar but darker background)