Add the keyboard event chaining.
The mp selection dialog is the only dialog which can benefit from it
so the only dialog which chains the keyboard events. (The window
itself is always chained eventhough there's no hotkey support yet.)

When the listbox has the focus and ends at the top item the next up
arrow is send to the textbox which selects from the history there. Not
sure whether this is great behaviour, but it can be changed by letting
the listbox always handle the up arrow.