Improve rendering support of blank lines...
...and fix bug #11925 (about blank label)

Now can render blank lines by using a double end-of-line in WML.

This is probably active in various places but mainly useful for
tooltips.  There is maybe newly visible blank lines, I can filter the
last one if needed but better clean the WML and keep the possibily to
use this.