correction to 2008-06-28T07:24:46Z!patrick_x99@hotmail.com...
...apply traits and modifications to attacks in unit::read()

note: this code needs an overhaul! The "real" max hitpoints, etc
should be stored in unit::write and WML changes to these should be
allowed to override the calculated values in unit::read. Also the _b_
variables in unit.cpp/hpp should be removed and reference the
unit_type directly when a "base" value is needed.