I fixed a warning for gcc 4.3.
I didn't change the semantics and the current way it works appears to
be reasonable, but I don't know enough about this section to be 100%
sure their wasn't a mistake that should have been corrected. It would
be nice if the person that changed it made sure that the semantics
were correct.