Make this AI require the unrestricted python environment.
It appears wail will likely require an unrestricted environmnet unless
python 2.6 becomes the default. This is because importing
"with_statement" seems to trigger an internal "__import__" which is
specifically verboten by the safe environment. Python 2.6 will have
"with" as a standard language feature which removes the requirement of
the __future__ import.

A hack/kludge may be possible in the future. For now, I'm not
wasting more time on safe.py which has already been a huge
time sink.