Fixed key-collision in ai_manager...
...which would cause an AI to be incorrectly reused by another
team. Added a method to allow the ai_manager to detect if an AI
requests management by the ai_manager. Lastly, added a hook to allow
the ai_manager to clean up AI resources prior to playsingle_controller
destruction.  Formula_ai now has method to inform ai_manager it wishes
to be managed.