Add some new transparent test images of Kitty.
These are used to test the new dialogs. They are writed in the unit
types with an undocumented API. This API will be documented once the
API is stable and also allow WML to modify the unit portraits, which
is not possible with the current API.

Note this doesn't interfere with the current profile key.
Code support will be added later.