Change the text storage type to std::string.
I thought the t_string had some magic to automatically change it's text
when the translation changes. That's not the case so no longer need it
to be a t_string. Also future changes to the text widgets will get
problems since I want to return references which might be temporaries
due to the conversions in t_string so rather avoid that as well.