Implement ~SECTION(x,y,w,h) imagepath function (feature / bug #12067).
There are some issues with this implementation - namely X and Y are
restricted positive because of limitations of the used cut_surface()
function- it cuts the width and height of the slice. This is not
possible to be overridden since the implementation uses memcpy
internappy. I need to write a cut_surface_ex() helper later.