This is the policy change that builds on my previous diff.
Turn the new-journey and new-battle dots red.  Addresses bug #12423:
Gold and silver journey dots are hard to see against the
parchment-toned mainmap.

By experiment, if the new dots are easy to see the older silver ones, attached, 
become much easier to notice.  The gold, though, was a real bitch to spot,
especially in mountains.