Zookeeper is right, it *is* silly that Galdrad issues his...
..."I am Galdrad. you will have to fight me to get any further!" and
then gets hammered flat almost instantly.  So let's power him up a
little.  With the number of orcs in play it's not going to actually
affect scenario balance noticeably.

(Yes, zookeeper is also right, he should be an Elvish Marshal.  But
loonycyborg is also also right that the AI doesn't use leadership
effectively, so the trait would be ineffective.)