Improved the map scrolling feature.
After 2008-11-21T21:05:43Z!kailoran@gmail.com fixed the scrolling outside the window, several
people noticed the threshold of 5 pixels is rather small. Added a
hidden preference named 'scroll_threshold' which is used to determine
the number of pixels. The threshold also defaults to 10 instead of 5,
more testing is needed to get a good default value.