Party fixed the scrollbar out of container drawing.
Ilor discovered the current language scrollbar can draw out of its
container. This commit fixes the drawing of items entirely out of the
container getting events (which caused the redraw). Items partly in the
container are still totally redrawn. A fix for that will be added later,
but is a bit more intrusive.