Add better draw support for invisible items.
Items in a scrollbar container which are not visible are no longer
redrawn even it they're dirty. (This was observed if a language which
was selected and scrolled out of sight still showed up when deselected.)
Items that are partly visible are still fully redrawn, this will be
addressed later.