Fixed a closing bug in the new dialogs.
The old style dialogs get dismissed on a mouse down, the next mouse up
event was most of the time handled by the new dialog. The new dialog
issued an easy_close event which closed the dialog. Moved the handler in
the mouse_up code (which already ignores mouse up events without the
button being down). The effect was percieved as the new dialog not
showing up.