Stricter check before sending events to widgets.
When a scrollbar button was clicked and that click disabled the button
since it reached the end of the range, the next mouse move event would
send a mouse leave event to that widget, which reenabled the widget.
Also added the check to the entering code (not sure whether it can be
triggered, but rather save than sorry).