Code cleanup step 1.
At the moment widget have an coordinate in the parent window and on the
screen. After the changes to the drawing engine the coordinate in the
parent window is not really needed anymore. So we can do several
refactoring and cleanup steps.

Remove the old coordinates and use the sceen coordinates everywhere.