Allow the message dialog to grow higher.
Kitty, Jetryl and zookeeper aren't too fond of the scrollbars in the
dialog if the text gets too long. Technically it's the easiest solution
to make things working in all cases. So we decided to let the dialog to
grow near the top so the scrollbars will pop up less frequently and
hopefully never.