make the reading-loop in network_worker check to see if the socket...
...has been interrupted. A very long read could make the disconnect()
time out before the reading was done, forcing a disconnect(force=true)
which cleans-up everything with little consideration to what is
actually going on. Now the reader thread will leave before the
time-out. Forced disconnect might still cause the crash, but there is
little reason for them to happen now. Sirp I would be gratefull for a
proofreading of that commit