Rewrote the rendering order of the engine.
The old engine had a hack to prevent glitches in the drawing order,
unfortunately this didn't work properly. At the FOSDEM looked with
boucman into the problem and invented a new rendering algorithm that
should work properly in all cases. (It have glitches with mountains, but
those are due to the fact that some parts of the mountain are drawn on
the foreground instead of at the background.)