Prepare 1.6 release notes
The template file has the same structure as the 1.4 release notes, but it is
still a work in progress and has no real content yet.

The Makefile used to build the release notes includes a few improvements: it
is now possible to build multiple .po and .html file simultaneously (see
make -j) and doesn't require updating the list of available translations.