The ~RC() image functor does not accept the special palette switch...
...~RC(palette1=palette2) syntax anymore. ~PAL(palette1>palette2)
should be used instead.

No idea what I was drinking when I decided to use ~RC(palette1=palette2)
instead of introducing a new functor (most likely I was thinking of the
overhead of the string comparison, but it's been reduced a lot since
then).

A wmllint test is missing, as well as backwards compatibility code. I
assume that this feature hasn't been used at all anyway.