Fix a possibly serious breakage of the AI:
The pathfinding 'distance' used by the AI was changed to 'moves left
at destination or 0 if unreachable in 1 turn', thus breaking some
target's rating.

This commit revert 2009-01-26T17:50:26Z!joerg.hinrichs@alice-dsl.de and 2009-01-25T23:48:17Z!joerg.hinrichs@alice-dsl.de which are not needed anymore.

I plan to clean this abuse of the name 'move_left' later, but first
need to test this simple fix for possible back-port to 1.6