Add a helper header for a small test.
Quoting the header:

This file is being used for a small experiment in which some private
functions of tgrid are no longer in tgrid but moved in a friend class with
static functions. The goal is to have less header recompilations, when
there's a need to add or remove a private function.
Also non-trivial functions like 'const foo& bar() const' and 'foo& bar()'
are wrapped in a template to avoid code duplication (for typing not for the
binary) to make maintenance easier.