Add parameter for the fake screen.
In order to make unit testing with different screen sizes possible it's
required to allow the setting of these values. Decided against doing it
in the constructor since I want to test on different screen sizes so
will rerun the function to get a different size. (Unit test code will be
committed later.)