Replace the layout algorithm.
The new algorithm seems to be at par with the old algorithm so make the
new one the algorithm used. The algorithm still needs some polishing but
the big problem is the old algorithm which doesn't know about the new
scrollbar mode. So once this algorithm gets a bit more testing in trunk
the old algorithm can be removed and this one can get polished.