The entire lower layer of save threading, in three lines of code!
Tested.  Congrats to YogiHH for making this so easy - I had been
planning to wait a couple of weeks but there's no need.

(Also, a tiny refactoring step using replace_space2underbar that was
clearly intended but missed.)

No actual behavior changes yet; that will take significant
modifications to the load- and save-game dialogs.  This, however, is
*all* the savefile-level infrastructure required.