Improve po-update to avoid nop updates.
The last change caused a lot of files be updated with the only change
the pot-update header. In the Wesnoth wiki there was a hint to use
msgattrib to cleanup the changes poedit makes. Using this on the file
seems to give the wanted output, proper files and no "dummy" updates.

Ivanovic please test this change.