Change the handling of double clicks.
The engine now always tries to generate a double click, if a double
click is generated it looks whether a widget wants double clicks when
building the event chain. Depending whether or not it supports a double
click it tries to send a double or single click.
After this decision it looks whether a handler is installed, but only
test the click or double click.