Address bug #14535, minor glitches in Delfador's Memoirs...
...(Mage Magister & Garard), by powering up the Mage Magister's staff
attack - for storyline reasons it should be better than a Great
Mage's.

This isn't as unbalancing as it looks; Delf relies on his fireballs
rather than staff, and the way the campaign is designed he's 
normally going to be pretty screwed if he goes into melee even
with this buff.