Allow content to notify about a resize.
At the moment when content of a scrollbar container needs a resize it
directly calls the window. If there are already scrollbars shown the
scrollbar container could handle the resize as well. Make that so.

When the scrollbars aren't shown the container might already be big
enough to handle the resize as well. This change is deemed too
intrusive for 1.8 and will be done post 1.8.