Fix a crash when destroying a window (MSVC only).
The chrash was caused by UB and only MSVC seemed to crash and our g++
users got unlucky never noticing the problem. The bug was spotted by
ilor and YogiHH.

Diagnosed and tested an earlier patch with YogiHH, this patch uses a
slightly different way, but should fix the problem as well. However it
would be nice if it gets more testing from a MSVC user.