Fix assertion failures in the tree view.
When opening or closing a node the sizes are now calculated better as
before and no longer cause assertions failures due to out of size.

Note the code still feels a bit hacky, but it seems to work properly so
rather commit now as rewriting it. The latter is scheduled to do later.