Fix a drawing glitch in the title screen.
The --new-widgets version using 800x480 with the German translation had
a drawing glitch*, where the background image wiped out a part of the
buttons on the right. This glitch is fixed. Another one remains the logo
is drawn over the area instead of below it, which will be fixed later.

*) other resolutions and translations might have the same problem, but
this was the easiest to reproduce for me.