Fixed broken attack predictions...
...when a unit is about to level up because of a kill.

Killing an enemy and losing hitpoints are not independent
events. (Quite the contrary in fact, the unit probably losed HPs
because it failed to kill its enemy.)

So advancement cannot be predicted by looking at the summaries after a
complex fight, since all the correlations have already been lost.

In other words, the previous code was completely incorrect. See
http://forums.wesnoth.org/viewtopic.php?f=4&t=30221 for an example.

As a side effect, this commit also improves predictions related to
slowness when leveling up after a complex fight.