Little bonus feature to check out terrain defense...
...for the units you select when it's not your turn.

Feedback needed, if people like it I'll do a bit of code cleanup,
otherwise this revision will be easy to revert.

This feature will be useful for the whiteboard's out-of-turn planning,
but I can always re-implement it specifically for the whiteboard move
arrows.

Note: the feature doesn't work while enemy units (at least AI units)
are moving/attacking.