Updated ana's implementation of get_bandwidth_stats()...
...to issue the string corresponding to network stats of the last hour
only. This is the original intention, as is visible in the
implementation in network.cpp, lines 961 through 969. Also added a
parameter to ana_components' method get_stats to select the time
period for which stats are requested and performed an analogous change
in the network_manager's method by the same name.