Stopped the Join button in the lobby from always being disabled.
The problem was that the server was sending the open game slots as a
string of the form "#/#", which lobby_data.cpp then tried to convert
to an int with: game["slots"].to_int()). The result of that was
obviously always zero.

I made the "slots" string to be written as only the number of open
slots, instead of using the form 'X slots'/'Y total'. Change it if you
like, but now everything seems to work as intended.