Reverted part of 2011-01-08T20:40:28Z!koraq@xs4all.nl, since it opened a major security breach...
...in Wesnoth.

Indeed, it caused the Lua engine to skip unwinding of its state in
presence of an exception, which opened the way to several attacks.

For instance, setting gc finalizers and then forcing GUI2 to throw an
exception would allow a multiplayer scenario to execute arbitrary code
on a remote client.