Fix unit_defense report generator doing unchecked map access (bug #17721)
This commit fixes report_unit_defense() to return an empty report
object if it detects that the game_display::displayed_unit_hex()
return value is not a valid map location. This problem is not really
recent, but fendrin's odd usage of the engine to display unit type
info on his recruit dialog unveiled this flaw in the worst manner
possible.