Added a new script: data/tools/terrain2wiki.py
Same as data/tools/terrain2wiki.rb except it works with current format
It also works with arguments, so you can choose source and destination
files

Still need some changes but it already works. Only feature that was in
ruby tool and not in this one is stats from. It will be added in a
future version.

mvt_alias and def_alias will also be added in a future version in the
alias column.

Another thing to not is that the tool does not parse terrains with an
'hidden=yes' because these shouldn't appear in a terrain reference.

Pep8 and Pylint results are good.