config: Treat VERSION specially when probing cache invalidation scenarios
It's not explained in the code at all, but I guess the point of
invalidating the cache when a non-empty or non-argless macro has been
defined is to make sure the cache was really automatically generated and
not tampered with by the user. In such case it should be safe to make an
exception for VERSION, the first engine-generated non-empty macro so
far.

I can't imagine a situation where this would break any other assumptions
at the moment, so I'll leave it to the testers.