WML preprocessor: implement #ifver, #ifnver MACRO <op> <version>
Currently, the specified MACRO is not expanded further, so if it
contains code inclusions with {} those will not be resolved.
Furthermore, MACRO should not take arguments. The whole idea is that it
should be defined to simple text such as "1.7.5" like Wesnoth does for
WESNOTH_VERSION since 2011-03-22T06:40:15Z!shadowm@wesnoth.org.