Ported the [allow_recruit] tag to lua and removed the c++ implementation.
Create an accessor for the help systems known units list for the lua
coder.  (used by allow_recruit)

Added an accessor for the new unit attribute "extra_recruit" for the lua coder.

Implemented [allow_extra_recruit], [dissallow_extra_recruit] and
[set_extra_recruit] by lua defined wml tags.

AI awareness is still to be done, the code locations are marked TODO
in unit.cpp.