Re-rationization of terrain names, part 1
Eleazar's "rationalization" of terrain names in 2010-10-22T20:51:37Z!me@jwbjerk.com,
2010-10-22T21:27:44Z!me@jwbjerk.com left quite a bunch of terrains with editor_name
attributes but no name attribute. While it may theoretically make
sense to inherit the name attribute from an aliased terrain, this
creates new problems of its own when a terrain is aliased to more than
one type (which one to choose for name inheritance?). In my opinion it
is best to solve this in WML and not add more layers (heh) of
convolutedness to the terrain engine.