Fix network communication between wesnoth and wesnothd, bug #18632
This is not a final fix.

wesnothd expects attributes to be sent in alphabetical order (sorted
by client) to reduce server load.  The new config attribute table is
no longer ordered.  The interim solution is to sort the attributes
before sending them. So multiplater client bear a sorted cost no paid
by single player clients.

A much better solution is to fix the server to not require sorting at
either end reducing network and CPU load.