Fixes #bug 18715
(Game hangs at loading screen after flushing caches and save index and
pressing F5)

Previously, when the .cache/wesnoth directory was deleted, while
wesnoth was running and then the cache was refreshed with F5 from the
startup screen, wesnoth hung.  The crash occured because wesnoth
didn't recreate the cache directory after startup, when it is
recovering from the cache not existing.  Now it makes a new cache
directory as needed.