Add a work-round for bug #18534.
The destructor of the class restored the moment of the units. When a
unit no longer exists it overwrites random memory. The issue has been
found by valgrind.

anonymissimus tested the work-around since he was able to reproduce the
issue easily, while on my system it went well for a long time.