Fix compilation with clang 2.9 / boost 1.47.
This applies a slightly modified version of the patch linked in
bug #18399. The changes are coding-style related.

The patch was intended to fix compilation with boost 1.47, but that
boost version is not yet in Debian so didn't test that part. Since it
does fix compilation with clang 2.9 Iapplied it, without waiting for
boost 1.47 to hit Debian.