Fixes crash when using fullscreen on Mac OS.
Fixes a problem when using the SDL nightly builds (post 1.2.14).

The direct blurring of the screen fails on OS X, since the surface is
not a neutral surface, so only testing for 32 bpp isn't enough. Exports
the neutral surface test from sdl_utils to have a proper test in the
canvas code.

This commit helps fixing bug #18319, but the speed of rendering is still
too slow.

Thanks to Alarantalara for testing the code on OS X.