Fixed bug #16571:
(Multiplayer login error with certain username or password formats)

The new username/password serialization format is as follows: @ <field
value> @ Old usernames/passwordswill be converted as necessary when
the user changes their credentials in-game. The user will to hit
#16571 once if they are using an affected username or password saved
by a previous version.

Unfortunately, the underlying cause is still around and this only
fixes this particular case by forcing new usernames and passwords to
be stored in a format that can only be interpreted as a string by the
config class. Dealing with it isn't feasible at this point, so I'll
file a separate bug for the general case and mark it as Postponed
later.