implement better faction detection (bug #19281)
Whenever default era is used the units should be placed for sure now,

this was IMHO not guaranteed since the recruits are internally a std::set,

so the first entry can change.

In case that one or both of the players have another faction than the
ones in the default era, none of the players gets the extra units.
Support for Khalifate can and should IMHO be added (given a suggested
set of units + locations).