Added unit sets for the "not a default faction" case composed by Doc Paterson.
If any faction is not one of the default factions (e.g. Khalifate)
it gets these units now instead of both factions getting no units
at all. If the other faction is a default one it still gets the set
composed for it in particular in the same game.

(bug #19281)

http://forums.wesnoth.org/viewtopic.php?p=518542&sid=73489ef47cba8b3a39ac219353f64613#p518542