Remove a backwards compat hack left over from previous commit:
2010-07-18T15:21:41Z!guillaume.melquiond@gmail.com

The hack causes problems now, since I want to load helper.lua previously
already and the comment says that it should be removed in 1.11.

I *think* the reason for the hack was to avoid loading wml-tags.lua twice
in case that a replay and/or (perhaps) savegame is loaded which was
created with a BfW version previously to 2010-07-18T15:21:41Z!guillaume.melquiond@gmail.com, 
since those replays/saves seemed to have recorded the commands for
loading the core lua files, which is no longer the case since back
then. Those replays/saves are no longer relevant, so everything hints
that I can safely remove this.