Adds new floating point emulation code.
This code can use several fixed point formats or a double to emulate
floating point code. The code is still work in progress. The intention
is to replace the current fixed point code and use it in more
areas. Hopefully this makes Wesnoth run snappier on devices with slow
floating point hardware, read the Pandora.

By default it uses doubles, so when used in existing code there should
be no differences, between the existing code and the new code.