addon/mg: Better fix for builds with non-GNU C++ STL than 2012-02-24T17:43:46Z!anonymissimus@arcor.de
The point of the at() member call was to throw an exception if the
access was invalid. By using find() without checking, we could end up
dereferencing the end iterator of the addons_list. Thus, a helper
function has been added now to use an assertion to protect against that
hypothetical situation early enough.