gui2/tgame_save_message, gui2/tgame_save_oos: Capture and grab textbox...
...values correctly, fixing bug #19562

The saved game filename textboxes in these dialogs appear to have been
mixed up with static labels in the C++ implementation of these dialogs,
so they weren't working as intended.

It was necessary to change some interfaces to accomodate this fix, but
it should be safe for backporting to 1.10 either way.