addon/mg: Set the Add-ons Manager dialog title to "Add-ons Manager"
Previously this was "Get Add-ons" or "Update Add-ons" depending on which
view was active (the latter was used for the upgradable add-ons view).
This distinction doesn't make much sense now that the dialog is
constructed in the same path and supports a third view mode. Thus, it
makes sense to use a generic catch-all title now instead.