gui2/teditor_settings: Use a cheaper,
...less CPU-intensive solution to refresh tiles after changing ToD settings

(Backported from trunk, 2012-03-21T08:23:30Z!shadowm@wesnoth.org and 2012-03-21T08:29:52Z!shadowm@wesnoth.org.)

Instead of refreshing the entire display, this solution simply
invalidates all gamemap tiles and requests them to be redrawn, saving a
lot of CPU time and making continuous operations (i.e. dragging one of
the sliders around) smoother. We still need to redraw the dialog,
though, and there's a hack involved to avoid a floating slice of it
appearing in the middle of the process.