fix $owner_side incorrectly set if capture event fired by unit movement
We are deleting the captured village from the set of owned villages of
the previous owner before querying which previous side owned it (and
giving it to the new side + firing the event), thus the result was
always "unowned".

fix for a bug reported in the forums:

http://forums.wesnoth.org/viewtopic.php?f=21&t=36569&p=525789#p525789

Backported from trunk 2012-04-07T16:43:03Z!anonymissimus@arcor.de.