Fix binding to a temporary.
The result of std::string::c_str() of a temporary std::string was stored
in an object to be used later. Changed the member to a std::string to
solve the issue.

NOTE since the code involved contains serveral OS specific #ifdef's
not all changes are tested.

Issue found by cppcheck; it choked on the code.