Add the find_at functions for the pane.
This allows the usage of the mouse in the pane. Also increases the size
of the pane to have a bit more space to move the mouse around.

Note the offset to the mouse movement and what it affects on the screen
is intentionally.

The code is used to experiment with a different approach of the
implementation of a listbox.