Move the widget fields to the widget class.
Some fields of the tbuilder_control are actually part of the twidget
class. These are now moved to the tbuilder_widget.

Some of these fields were also used in the tbuilder_grid and are moved
likewise. This means the tbuilder_grid now has some extra fields for
debugging. (`Inherited' from the tbuilder_control.)