addon/mg: Horrible, horrible hack to ensure the add-ons list (GUI1)...
...is naturally sorted by titles rather than case-sensitive ids

This is quite a hack and it's only here in case I don't figure out a
better solution later before 1.11.0 (or get hit by a truck). It works
and that's all that matters here nowadays, anyway.

The problem this solves is basically the following:

 * The internal add-ons list is sorted by id (case-sensitively)

 * The add-ons list in GUI1 is unsorted by default, following the row
   insertion order

 * The row insertion order is that of the internal add-ons list iterator order

 * Sorting rows by title isn't possible because of the Publish/Delete entries

 * Many add-ons have titles that don't reflect their ids at all, not to
   mention that users obviously would prefer case-insensitive sorting.

A quite better solution would be to have the add-ons list take care of
all this by itself, but I haven't figured out a way yet short of
implementing a whole custom class for it (which could actually turn out
to be more efficient by getting rid of the id duplication) with
specialized iterators.