addon/mg: Rearrange list columns in roughly descending order of importance
With translations, running out of space due to the Type column and the
status line (which expands the Title column) is an inevitability. Thus
we need another compromise.

Type has to go last since it can be the longest due to the non-obvious
acronym usage (e.g. "MP scenario"). A later commit will introduce more
advanced filtering to make up for it when it goes missing due to space
constraints.

It should be noted that this is mostly relevant only for horizontal
resolutions < 1024.

The old column order:

  AM: Icon Name Version Author Type Downloads Size
  DP: Icon Name Version Author Type Size
  UA: Icon Name OldVersion NewVersion Author Size

The new order:

  AM: Icon Name Version Author Size Downloads Type
  DP: Icon Name Version Author Size Type
  UA: Icon Name OldVersion NewVersion Author Size

(AM being the Add-ons Manager; DP, the Dependency Prompt; and UA, the
Upgradable Add-ons view.)

The most important difference is that Size goes always before Downloads
and after Author now since people with low resolution displays are very
likely using services with poor/restricted Internet traffic; making sure
they can see the Size without reading the Description dialog seems only
logical.

An upcoming commit with make the gui2::taddon_description dialog display
all of the information just in case.