Switch the default gui2 render engine.
In gui2 there are two render engines; one using the widget's position to
render and the other based on the requested coordinates. The later is
used in the new widgets (pane, viewport and matrix) exclusively, meaning
these widgets don't work with the old engine.

The new one has been tested a while under --new-widgets and seems to
have no glitches so make it the default.

The old code hasn't been removed since doing so will force a huge
recompilation and it's desired to quickly switch back in case there
has been a regression.

The code is used to experiment with a different approach of the
implementation of a listbox.