Adds the ENABLE_SHARED_LIBRARIES CMake option.
The use of shared libraries makes compilation debug versions faster but
results in extra shared libraries. For installation this is not
practical since the libraries need to be given SONAMES and kept in sync.
Therefore installation is not supported for this option and will
probably fail.

The statistics for the change are:

Static build:
  - Normal:
      real 20m51.604s
      user 19m47.838s
      sys  1m5.100s
      size 94MB
  - Debug:
      real 24m48.577s
      user 23m3.114s
      sys  1m22.305s
      size 1.9GB

Shared build:
  - Normal:
      real 21m5.141s
      user 19m56.879s
      sys  1m7.692s
      size 29MB
  - Debug:
      real 24m24.239s
      user 22m55.278s
      sys  1m20.193s
      size 596MB

So the build times don't change that much but the sizes do. When
rebuilding after changing a file the build times start to improve for
the shared build, but not timing done.