Do not always clear end_turn_ and hold_position_ in unit::set_movement().
This allows the function to be called without side-effects, so it can
be used as a simple setter (as the name suggests). Since the previous
behavior was to have side-effects (which is bad style for a default),
this commit introduces a non-optional parameter, ensuring that I've
looked at each place where this function is called.