Implement [effect] apply_to=experience (bug #16747)
This implements 'increase' and 'set' attributes with similar semantics
to apply_to=hitpoints effects. The 'should_advance' attribute from the
FR is not implemented at this point since the unit class doesn't really
know how to present the user with an advancement prompt, and it
shouldn't really be its business to do so either.

As-is, using [effect] to set a unit's current XP above its maximum
simply won't trigger any advancements. It should be possible to work
around this limitation with [unstore_unit] (which is a real WML action,
unlike [effect]) and other tricks.