Refactor weapon special contexts.
This allows some simplification both when the context is set and (in
one spot) when the context is used.

The semantics for the location parameters to set_specials_context() have
changed; attempts to use the old version will give a compiler error as the
function signatures have also changed (all current calls have been updated).