Correcting attack predictions for swarm...
...when there are previous combats to account for.

(It passes the attack simulator check.)

The basic problem with the old system is that it treated the number of
strikes as an independent variable, even though it depends on the
number of hit points.

There is some additional overhead with this method, but I optimized
some things for the common case (no swarm), and benchmarking indicated
a negligible 1% increase in run time when no units had swarm.