Instead of removing prestart events...
...(with a method that fails if an event's name is, e.g.,
"prestart,turn_end"), simply do not fire those events when loading a
replay (analogous to not firing those events when loading a game
normally).

This reverts most of 2012-11-07T19:28:44Z!therattigan@gmail.com, replacing it with a simpler fix to
bug #20272.  This also handles the additional case of start events.