When creating [replay_start],
...have the snapshot overwrite the scenario config, rather than the
other way around.

This is important for the RNG as the snapshot includes how often the
RNG was invoked in prestart and start evnets, while the scenario's
config has the RNG recorded as having never been invoked (naturally
enough).  It might solve some other replay issues as well. (Not all
replay issues are well-documented.)