Restore 2010-07-20T10:46:41Z!guillaume.melquiond@gmail.com,
by removing 2010-07-23T05:51:17Z!upthorn@gmail.com 
and the follow-ups 2010-07-23T06:31:24Z!guillaume.melquiond@gmail.com and 2010-07-30T21:49:33Z!koraq@xs4all.nl.

In theory, the underlying bug has been fixed, so no need for the special
code for MSC. If this is not the case (the symptom would be "the windows
executable crashes immediately before loading the main menu"), feel free
to revert this.