gui1: Rework textbox event forwarding  to eliminate duplicate...
...(actually quadruplicate) events.  Changed from 2012-12-23T06:10:46Z!shadowm@wesnoth.org.

Due to the way focusing works, the focus switch in the middle of the
forwarding action would create a duplicate event for the origin textbox,
resulting in a duplicate event for the target (a total of four events to
process).

Instead, make the implementation of gui::textbox::handle_event()
forwarding-aware, and provide a wrapper for the events engine
implementation to use by default.