Make unit_type_data::build_unit_type() public,
...and update various parts of the code to call it (instead of using
unit_type_data::find() to "find" a known unit_type).

There was one call in the editor's code where I am unsure if there are
supposed to be side effects other than building the unit_type. So I
left the code as-is and just added a comment for someone else to
evaluate later.