Set up multiplayer games started from the commandline...
...equivalently to lobby games.

This means that parameters now use the correct default values,
scenarios using random maps work and side leaders always have unique
IDs.  Scenario and side parameters are defined as follows, in order of
increasing priority:

  Wesnoth defaults --> map settings --> commandline-specified values

All modifications are in a code path separate from lobby games (as it
was before), so lobby games are not affected.

This fixes bugs #19853, #19883, #19895 and #20009.