Reduce the data stored in the undo stack.
Most actions had caused a unit to be stored in the stack, but the stack
does not really need all the data from a unit. Now it will store only
those parts it needs, which might have a significant impact in
scenarios with WML-heavy units. There is also a reduction in the size
of mid-turn saves (~9% in one test case using NR:Showdown).