Added timestamps in Chatlog
It was a little odd that timestamps checkbox in Preferences affected
only regular chat messages, but not the chatlog, so I decided to add
timestamps into the chatlog.

I hate to use `const_cast` to add a timestamp to a received message, so
if someone knows a better solution, don't hesitate to mention it.